API changes needed for BTLE work
