+++ webauthn-authenticator-rs/src/cable/tunnel.rs
@@ -387,7 +387,7 @@ impl Token for Tunnel {
         Ok(data)
     }
 
-    fn cancel(&self) -> Result<(), WebauthnCError> {
+    async fn cancel(&self) -> Result<(), WebauthnCError> {
         // There is no way to cancel transactions without closing in caBLE
         Ok(())
     }
